% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTdeconv.R
\name{CTdeconv}
\alias{CTdeconv}
\title{Cell types deconvolution}
\usage{
CTdeconv(mix, RNAseq = F, filename = NULL)
}
\arguments{
\item{mix}{A matrix (n Genes * m Samples) of genes expression from bulk samples
. It can be a string which specifies the path of a expression 'tsv' file. The rownames
must be gene symbols and the colnames are the sample id.}

\item{RNAseq}{Logical. if TURE, the expression data is RNA-Seq data rather than
Macroarray data, and the quantile normalization will not be done during Cibersort
analysis process. The default is FALSE.}

\item{filename}{A string indecates the path of the file to be saved. If NULL (default),
no file will be saved.}
}
\value{
A matrix (m Samples * 6 Celltypes). It provides the proprotions of
six cell types in the bulk samples. Note that the proportion is relative since we
conduct a normalization to make the sum of proportions in each sample be one.
}
\description{
This function integrates deconvolution results of three agrithom-signature
combinations which show best performance in our benchmark datasets and provides
the relative proportions of six immune cell types in mixture samples.
}
\examples{
res=CTdeconv(mix)
}
