% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bm.R
\name{fbm}
\alias{fbm}
\title{Generate a time series of fractional Brownian motion.}
\usage{
fbm(hurst = 0.7, n = 100)
}
\arguments{
\item{hurst}{ggg}

\item{n}{xiergo}
}
\value{
haha
bbb
ccc
}
\description{
This function generatea a time series of one dimension fractional Brownian motion.
adapted from http://www.mathworks.com.au/matlabcentral/fileexchange/38935-fractional-brownian-motion-generator .
}
\details{
I am xieyuhao
and this is my test

good night
hai

woshixiaohai
}
\examples{
fbm()
plot(fbm())
d <- fbm(hurst=0.2, n=1000)
plot(d)
}
