% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{lm22}
\alias{lm22}
\title{LM22 signature matrix}
\format{A matrix.}
\source{

}
\usage{

}
\description{
Signatures of 22 celltypes. It is provided by Cibersort.
}
\examples{

}
\references{

}
\keyword{datasets}
