% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{lm6}
\alias{lm6}
\title{LM6 signature matrix}
\format{A matrix.}
\source{

}
\usage{
data(lm6)
}
\description{
Signatures of six celltypes. It is provided by Cibersort.
}
\examples{

}
\references{

}
\keyword{datasets}
