% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mydata.R
\docType{data}
\name{mix}
\alias{mix}
\title{Mixture expression matrix.}
\format{A matrix.}
\source{

}
\usage{
data(mix)
}
\description{
The expression profiles of 5 mixtures. It is just a example data.
}
\examples{

}
\references{

}
\keyword{datasets}
